+++ changelog.md
@@ -1,3 +1,21 @@
+## v0.6.3 [view commit logs](https://github.com/marionettejs/backbone.syphon/compare/v0.6.2...v0.6.3)
+
+### Fixes
+
+* Update Backbone version range
+
+## v0.6.2 [view commit logs](https://github.com/marionettejs/backbone.syphon/compare/v0.6.1...v0.6.2)
+
+### Fixes
+
+* Fix botched release :(
+
+## v0.6.1 [view commit logs](https://github.com/marionettejs/backbone.syphon/compare/v0.6.0...v0.6.1)
+
+### Fixes
+
+* Fix the inability to find nested inputs in a form.
+
 ## v0.6.0 [view commit logs](https://github.com/marionettejs/backbone.syphon/compare/v0.5.1...v0.6.1)
 
 ### Fixes